fix: use isoWeek instead of weekOfYear, and rename file for active staking helpers
