+++ src/rtk/features/activeStaking/rewardHistorySlice.ts
@@ -1,5 +1,5 @@
 import { createEntityAdapter, createSlice } from '@reduxjs/toolkit'
-import { getRewardHistory } from 'src/components/utils/datahub/super-likes'
+import { getRewardHistory } from 'src/components/utils/datahub/active-staking'
 import { RootState } from 'src/rtk/app/rootReducer'
 import { createSimpleFetchWrapper } from 'src/rtk/app/wrappers'
